add contact group

+++ app/src/main/java/com/ae/apps/randomcontact/room/repositories/ContactGroupRepository.kt
@@ -17,7 +17,10 @@ class ContactGroupRepository private constructor(private val dao:ContactGroupDao
 
     fun getAllContactGroups() = dao.getAll()
 
-    fun findContactGroupById(contactId:String) = dao.findContactGroupById(contactId)
+    // Id is stored in the database as an Integer inorder to apply AutoIncrement
+    // But it is stored in Preferences as a String, so inorder to maintain backwards compatibility,
+    // we are converting to an int here
+    fun findContactGroupById(contactId:String) = dao.findContactGroupById(contactId.toInt())
 
     fun createContactGroup(contactGroup:ContactGroup) = dao.insert(contactGroup)
